+++ core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.7
 
+- Add polynomial multiproof features on core
 - Changed the default AutoNat `only_global_ips` from `false` to `true` so the servers (and clients) discard AutoNat probes
   from the private-only peers
 - Updated the default AutoNat configuration with increased timeouts